SplitDebug: Install debug symbols in same location, strip them as well.
Installing the debug symbols into the same location as the binaries themselves
causes GDB to always find them. As such we no longer need to keep the code
itself inside the .debug files and can strip that away, resulting in a cleaner
install. If any systems require the debug symbols to reside in some different
location, that should be handled by packaging (which it is currently not).